# app/api/swagger.yaml
openapi: 3.0.0
info:
  title: MealTime API
  version: 2.0.0
  description: |
    API completa para gerenciamento de alimentação de gatos
    
    **VERSÕES DISPONÍVEIS**:
    - **V2 (Recomendado)**: `/api/v2/*` - Autenticação JWT/Session, respostas padronizadas
    - **V1 (Deprecated)**: `/api/*` - Será removido em 2025-07-28
    
    **AUTENTICAÇÃO V2**:
    - Mobile: JWT via header `Authorization: Bearer <token>`
    - Web: Supabase Session via cookies
  contact:
    name: MealTime Support
    email: support@mealtime.app
servers:
  - url: http://localhost:3000/api/v2
    description: Servidor V2 (desenvolvimento) - RECOMENDADO
  - url: https://mealtime.app/api/v2
    description: Servidor V2 (produção) - RECOMENDADO
  - url: http://localhost:3000/api
    description: Servidor V1 (desenvolvimento) - DEPRECATED
  - url: https://mealtime.app/api
    description: Servidor V1 (produção) - DEPRECATED
...
paths:
...
  /v2/cats:
    get:
      tags: [V2 - Cats]
      summary: 🆕 Listar gatos (V2)
      description: Retorna lista de gatos do household do usuário autenticado
      security:
        - bearerAuth: []
      parameters:
        - name: householdId
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por household específico (opcional)
      responses:
        '200':
          description: Lista de gatos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Response'
              examples:
                success:
                  value:
                    success: true
                    data:
                      - id: "uuid"
                        name: "Miau"
                        photo_url: "https://..."
                        birth_date: "2020-01-15"
                        weight: "4.5"
                        household_id: "uuid"
                        owner_id: "uuid"
                        created_at: "2025-01-28T10:00:00Z"
                        updated_at: "2025-01-28T10:00:00Z"
                    count: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags: [V2 - Cats]
      summary: 🆕 Criar gato (V2)
      description: Adiciona um novo gato ao household
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, householdId]
              properties:
                name:
                  type: string
                householdId:
                  type: string
                  format: uuid
                photoUrl:
                  type: string
                birthdate:
                  type: string
                  format: date-time
                weight:
                  type: number
                feeding_interval:
                  type: integer
                  minimum: 1
                  maximum: 24
      responses:
        '201':
          description: Gato criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Response'
...
  /v2/goals:
    get:
      tags: [V2 - Goals]
      summary: 🆕 Listar metas de peso (V2)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de metas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Response'
    
    post:
      tags: [V2 - Goals]
      summary: 🆕 Criar meta de peso (V2)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cat_id, goal_name, target_weight]
              properties:
                cat_id:
                  type: string
                  format: uuid
                goal_name:
                  type: string
                target_weight:
                  type: number
                unit:
                  type: string
                  enum: [kg, lbs]
      responses:
        '201':
          description: Meta criada

  # ==================== V2 - SCHEDULES (RECOMENDADO) ====================
...
  /v2/households/{id}/cats:
    get:
      tags: [V2 - Households]
      summary: 🆕 Listar gatos do household (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de gatos
    
    post:
      tags: [V2 - Households]
      summary: 🆕 Adicionar gato ao household (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Gato adicionado
...
  /cats:
    get:
      tags: [Cats]
      summary: ⚠️ Listar gatos (V1 - DEPRECATED)
      description: |
        **DEPRECATED**: Use `/v2/cats` em vez disso.
        Esta rota será removida em 2025-07-28.
        
        Retorna lista de gatos do usuário
      deprecated: true
      parameters:
        - name: householdId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de gatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cat'
    
    post:
      tags: [Cats]
      summary: Criar gato
      description: Adiciona um novo gato
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatInput'
      responses:
        '201':
          description: Gato criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
  /cats/{catId}:
    get:
      tags: [Cats]
      summary: Obter gato por ID
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do gato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
    
    put:
      tags: [Cats]
      summary: Atualizar gato
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatInput'
      responses:
        '200':
          description: Gato atualizado
    
    delete:
      tags: [Cats]
      summary: Deletar gato
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Gato deletado com sucesso
...
  /households/{id}/cats:
    get:
      tags: [Households]
      summary: Gatos do domicílio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de gatos do domicílio
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtido via POST /api/auth/mobile
  responses:
    Unauthorized:
      description: Não autorizado - Token ausente ou inválido
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Token de autorização não fornecido ou sessão inválida"
    Forbidden:
      description: Acesso negado - Usuário não tem permissão
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Acesso negado ao household"
  schemas:
    V2Response:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indica se a requisição foi bem-sucedida
        data:
          type: object
          description: Dados retornados (quando success=true)
        count:
          type: integer
          description: Número de itens retornados (para listas)
        error:
          type: string
          description: Mensagem de erro (quando success=false)
        details:
          type: object
          description: Detalhes adicionais do erro (quando aplicável)
      example:
        success: true
        data:
          id: "uuid"
          name: "Exemplo"
        count: 1
    Cat:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        photoUrl:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        householdId:
          type: string
        ownerId:
          type: string
        portionSize:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatInput:
      type: object
      required: [name, householdId]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        householdId:
          type: string
        portionSize:
          type: number
...
security:
  - bearerAuth: []